chore: refactor repo structure
