+++ Makefile
@@ -7,22 +7,22 @@ help:
 	@echo ""
 	@echo "make lint                 	run golangci-lint"
 	@echo "make test                 	run  tests"
-	@echo "make format             		format files"
+	@echo "make format             	format files"
 	@echo "make build             		build files"
 	@echo "make setup/git/hooks      	setup git hooks"
 .PHONY: help
 
 run:
-	go run ./cmd/charmil
+	go run ./cli/cmd/charmil
 
 build:
-	go build ./cmd/charmil
+	go build ./cli/cmd/charmil
 
 install:
-	go install ./cmd/charmil
+	go install ./cli/cmd/charmil
 
 build/tags:
-	go build -tags ${TAGS} ./cmd/charmil
+	go build -tags ${TAGS} ./cli/cmd/charmil
 
 test:
 	go test ./...
@@ -30,14 +30,13 @@ test:
 # Requires golangci-lint to be installed @ $(go env GOPATH)/bin/golangci-lint
 # https://golangci-lint.run/usage/install/
 lint:
-	golangci-lint run cmd/... pkg/... 
+	golangci-lint run ./...
 .PHONY: lint
 
 
 # clean up code and dependencies
 format:
 	@go mod tidy
-
 	@gofmt -w `find . -type f -name '*.go'`
 .PHONY: format
