+++ cli/internal/cmd/crud/crud.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/aerogear/charmil/cli/internal/template/crud"
 	"github.com/aerogear/charmil/core/factory"
-	"github.com/aerogear/charmil/pkg/template/crud"
 	"github.com/spf13/cobra"
 )
