+++ cli/internal/template/crud/crud.en.yaml
