+++ cli/internal/template/crud/delete.go
