+++ cli/internal/template/crud/describe.go
