+++ cli/internal/template/crud/list.go
