+++ cli/internal/template/crud/tmpl.go
